Донастройка логгера
Вроде как настроили покруче логгер, теперь и логи покрасивее и в сервере убрали require('dotenv').config();, так теперь типа это Не нужно require('dotenv').config(); — его уже загрузили через -r dotenv/config
