+++ package.json
@@ -2,7 +2,7 @@
   "name": "match-predictions",
   "version": "1.0.0",
   "description": "Сайт прогнозов на матчи с авторизацией на Node.js",
-  "main": "NODE_PATH=. nodemon --experimental-specifier-resolution=node server.js",
+  "main": "node --experimental-specifier-resolution=node -r dotenv/config server.js",
   "scripts": {
     "start": "node server.js",
     "dev": "nodemon server.js"
